Added Author to Post
Now, we can know who wrote the post with the Author label.

Also, I created a new displayPosts.html.twig, because on the home page and the search post result page, there is the exactly same displaying result format of posts. So I included this template of posts displaying in the two twig file (home and search) and pass the posts variable.