Updated PostController api v1, return xml or json for posts and post/id
I updated the api/v1/PostController to respect all the REST convention. For the getPostsAction and getPostAction, both return JSON or XML response.

Default is JSON  but if the user send an HTTP header "Accept" with "application/xml", the API return an XML format.

And I also add a name to the 2 routes